refactor: update StationData structure and improve timestamp handling

+++ internal/netatmo/format.go
@@ -57,12 +57,12 @@ func FormatBattery(percent int) string {
 	}
 }
 
-func FormatRelativeTime(timestamp int) string {
+func FormatRelativeTime(timestamp int64) string {
 	if timestamp == 0 {
 		return "unknown"
 	}
 
-	t := time.Unix(int64(timestamp), 0)
+	t := time.Unix(timestamp, 0)
 	diff := time.Since(t)
 
 	switch {
